+++ app/models/user.py
@@ -1,7 +1,7 @@
 from .db import db
 from werkzeug.security import generate_password_hash, check_password_hash
 from flask_login import UserMixin
-
+from .party import users_parties
 
 class User(db.Model, UserMixin):
     __tablename__ = 'users'
@@ -52,28 +52,3 @@ class User(db.Model, UserMixin):
             'job': self.job.name,
             'role': self.job.role.name
         }
-
-class Party(db.Model):
-    __tablename__ = 'parties'
-
-    #columns
-    id = db.Column(db.Integer, primary_key= True)
-    postId = db.Column(db.Integer, db.ForeignKey('posts.id'))
-    ownerId = db.Column(db.Integer, db.ForeignKey('users.id'))
-    createdAt = db.Column(db.DateTime , nullable= False)
-    updatedAt = db.Column(db.DateTime , nullable= False)
-
-    #relationships
-    users = db.relationship('User', secondary='users_parties', back_populates= 'parties')
-
-
-users_parties = db.Table(
-    "users_parties",
-    db.Column('id',db.Integer, primary_key=True),
-    db.Column(
-        'userId', db.Integer, db.ForeignKey("users.id")
-    ),
-    db.Column(
-        'partyId', db.Integer, db.ForeignKey('parties.id')
-    )
-)
